+++ lib/maze/configuration.rb
@@ -9,6 +9,7 @@ module Maze
       self.receive_no_requests_wait = 30
       self.receive_requests_wait = 30
       self.enforce_bugsnag_integrity = true
+      self.captured_invalid_requests = Set[:errors, :sessions, :builds, :uploads, :sourcemaps]
     end
 
     #
@@ -55,6 +56,9 @@ module Maze
 
     # Enables bugsnag reporting
     attr_accessor :enable_bugsnag
+
+    # The server endpoints for which invalid requests should be captured and cause tests to fail
+    attr_accessor :captured_invalid_requests
 
     #
     # General appium configuration
@@ -124,6 +128,9 @@ module Maze
 
     # URI of the test-management service
     attr_accessor :tms_uri
+
+    # Access token for the test-management service
+    attr_accessor :tms_token
 
     #
     # Local testing specific configuration