+++ lib/maze/option.rb
@@ -30,6 +30,7 @@ module Maze
     OS = 'os'
     OS_VERSION = 'os-version'
     APP_BUNDLE_ID = 'app-bundle-id'
+    LIST_DEVICES = 'list-devices'
 
     # CrossBrowserTesting/Bitbar options
     SB_LOCAL = 'sb-local'
@@ -40,7 +41,6 @@ module Maze
     # Sauce Labs-only options
     SL_LOCAL = 'sl-local'
 
-    # BitBar-only options
     TMS_URI = 'tms-uri'
 
     # Local-only options