+++ lib/maze/option/processor.rb
@@ -21,6 +21,8 @@ module Maze
           # General options
           config.enable_retries = options[Maze::Option::ENABLE_RETRIES]
           config.enable_bugsnag = options[Maze::Option::ENABLE_BUGSNAG]
+          config.tms_uri = options[Maze::Option::TMS_URI]
+          config.tms_token = options[Maze::Option::TMS_TOKEN]
 
           # Document server options
           config.document_server_root = options[Maze::Option::DS_ROOT]