+++ test/fixtures/android-appium-test/features/steps/steps.rb
@@ -9,3 +9,7 @@ end
 When('I run the next command') do
   step 'I click the element "run_command"'
 end
+
+Then('Maze Runner reports the current platform as {string}') do |platform|
+  Maze.check.equal(platform, Maze::Helper.get_current_platform)
+end